fix: technologyCode in dataprocessings
We had the same issue in the usage module.
We can reuse the same function to solve the issue.
